pyproject.toml
[project]
name = "table-agent"
version = "0.1.0"
description = "多格式文件语义处理框架 - 支持 PDF/PPT/Word/Excel/CSV/MP4"
requires-python = ">=3.11"
dependencies = [
    "markitdown[all]",
    "openai>=1.0",
    "opencv-python",
    "pydantic>=2.0",
    "pyyaml",
    "python-frontmatter",
]

[project.scripts]
table-agent = "table_agent.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/table_agent"]
